Render at the display refresh rate with interpolated motion
The simulation still runs at its classic deterministic 30Hz — replays,
lockstep networking, and game speed are unchanged — but rendering is now
decoupled from it. Between simulation steps, objects, shots, and the
zoom camera are drawn at positions interpolated between the previous and
current step (with torus-aware wrapping at the playground edges), at the
refresh rate of the display showing the window: 120fps on ProMotion
displays, 60fps on standard ones, re-queried when the window changes
displays. Sprites are now blitted at subpixel positions for smoothness.

The shield/thrust animation and the score-based extra-life award moved
from the draw path into the simulation so drawing has no side effects.

A new "Smooth motion" checkbox in Game Options (InterpolateMotion pref,
default on) falls back to the original one-draw-per-step behavior.

Verified with full-length replay playback: existing recordings play back
bit-exact (identical final scores, no RNG consistency errors).

+++ screenlib/SDL_FrameBuf.h
@@ -109,12 +109,12 @@ public:
 	/* Blit and update routines */
 	void QueueBlit(SDL_Texture *src,
 			int srcx, int srcy, int srcw, int srch,
-			int dstx, int dsty, int dstw, int dsth, clipval do_clip, float angle = 0.0f);
-	void QueueBlit(SDL_Texture *src, int x, int y, int w, int h, clipval do_clip, float angle = 0.0f) {
+			float dstx, float dsty, float dstw, float dsth, clipval do_clip, float angle = 0.0f);
+	void QueueBlit(SDL_Texture *src, float x, float y, float w, float h, clipval do_clip, float angle = 0.0f) {
 		QueueBlit(src, 0, 0, src->w, src->h, x, y, w, h, do_clip, angle);
 	}
 	void QueueBlit(SDL_Texture *src, int x, int y, clipval do_clip, float angle = 0.0f) {
-		QueueBlit(src, 0, 0, src->w, src->h, x, y, src->w, src->h, do_clip, angle);
+		QueueBlit(src, 0, 0, src->w, src->h, (float)x, (float)y, (float)src->w, (float)src->h, do_clip, angle);
 	}
 	void StretchBlit(const SDL_Rect *dstrect, SDL_Texture *src, const SDL_Rect *srcrect);
